new feature: add custom messages on screen
New feature allows to create custom messages in pattern file with M_here your message. The start coloumn is used to define the time when the message will appear on the display. With the stimulation time, a time for the corresponding countdown can be defined. Up to 10 messages can be defined in a pattern file. At one time 2 messages can be displayed simultanously.
Additionally, update display will now only be called every 100ms instead of every CPU cylce.