Adapt Platformio.ini and code to be able to compile LITOS source code for both board revisions, Rev 2 and Rev4.
The board revisions 2 and 4 are distinguished by different button mappings. Further, some of the Rev2 boards had problems with a spontanous triggering of the quit (green) button. Thus, the exit of the program was moved to the red button for Rev2 boards.
Pin definitions for the baords can be found in button.h