LITOS Release 2.0 with hardware revision 4
Compared to the prerelease of revision 4, C5 had been removed. It seems that the capacitors attached to the buttons of RST and BOOT lock the ESP32 into the bootloader when doing a cold start as both pins are pulled to GND during the initial charging phase of the caps. By removing C5 (which serves only as a debouncing CAP for the BOOT pin we do not have any problems anymore when cold starting LITOS.
Further the board was shortened by 1 mm, as the linked micro-USB connector could not be soldered flat on the board, as some overhang was needed for the micro-USB connector.
To ensure compatibility with LITOS Rev2 and Rev3, we added a LITOS_Rev2_Rev3 build environment in the platform.ini file, for more information see the hardware readme.
An optional 3D printed case for LITOS Rev4 had also been added, you can choose between a case with ventilation vents and a solid case that protects LITOS more from dust